SW-CR-348 Convert SingleScanSettings column to an int.

+++ Code/ShipWorks.Data.Model/EntityInterfaces/IUserSettingsEntity.cs
@@ -120,9 +120,9 @@ namespace ShipWorks.Data.Model.EntityInterfaces
         /// <summary> The SingleScanSettings property of the Entity UserSettings<br/><br/>
         /// </summary>
         /// <remarks>Mapped on table field: "UserSettings"."SingleScanSettings"<br/>
-        /// Table field type characteristics (type, precision, scale, length): TinyInt, 3, 0, 0<br/>
+        /// Table field type characteristics (type, precision, scale, length): Int, 10, 0, 0<br/>
         /// Table field behavior characteristics (is nullable, is PK, is identity): false, false, false</remarks>
-        System.Byte SingleScanSettings { get; }
+        System.Int32 SingleScanSettings { get; }
         
         IUserEntity User { get; }
